refactor: 필요없는 메서드 제거 및 패키지 구조 변경
