+++ backend/src/main/java/botobo/core/common/domain/BaseEntity.java
@@ -1,4 +1,4 @@
-package botobo.core.quiz.domain;
+package botobo.core.common.domain;
 
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
