+++ backend/src/main/java/botobo/core/common/exception/GlobalControllerAdvice.java
@@ -1,4 +1,4 @@
-package botobo.core.exception;
+package botobo.core.common.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.support.DefaultMessageSourceResolvable;
