+++ backend/src/main/java/botobo/core/common/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package botobo.core.exception;
+package botobo.core.common.exception;
 
 public class UnauthorizedException extends RuntimeException {
