+++ backend/src/main/java/botobo/core/quiz/exception/WorkbookNotFoundException.java
@@ -1,6 +1,6 @@
 package botobo.core.quiz.exception;
 
-import botobo.core.exception.NotFoundException;
+import botobo.core.common.exception.NotFoundException;
 
 public class WorkbookNotFoundException extends NotFoundException {
     public WorkbookNotFoundException() {
