+++ backend/src/test/java/botobo/core/admin/AdminAcceptanceTest.java
@@ -5,7 +5,7 @@ import botobo.core.admin.dto.AdminCardResponse;
 import botobo.core.admin.dto.AdminWorkbookRequest;
 import botobo.core.admin.dto.AdminWorkbookResponse;
 import botobo.core.auth.AuthAcceptanceTest;
-import botobo.core.exception.ErrorResponse;
+import botobo.core.common.exception.ErrorResponse;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
