+++ backend/src/test/java/botobo/core/quiz/QuizAcceptanceTest.java
@@ -3,7 +3,7 @@ package botobo.core.quiz;
 import botobo.core.admin.dto.AdminCardRequest;
 import botobo.core.admin.dto.AdminWorkbookRequest;
 import botobo.core.auth.AuthAcceptanceTest;
-import botobo.core.exception.ErrorResponse;
+import botobo.core.common.exception.ErrorResponse;
 import botobo.core.quiz.dto.QuizRequest;
 import botobo.core.quiz.dto.QuizResponse;
 import io.restassured.RestAssured;
