+++ backend/src/test/java/botobo/core/quiz/domain/card/CardTest.java
@@ -1,4 +1,4 @@
-package botobo.core.quiz.domain;
+package botobo.core.quiz.domain.card;
 
 import botobo.core.quiz.domain.card.Card;
 import botobo.core.quiz.domain.workbook.Workbook;
