+++ backend/src/test/java/botobo/core/quiz/domain/workbook/WorkbookRepositoryTest.java
@@ -1,7 +1,5 @@
-package botobo.core.quiz.domain;
+package botobo.core.quiz.domain.workbook;
 
-import botobo.core.quiz.domain.workbook.Workbook;
-import botobo.core.quiz.domain.workbook.WorkbookRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -41,20 +39,6 @@ public class WorkbookRepositoryTest {
         assertThat(savedWorkbook.getUpdatedAt()).isNotNull();
         testEntityManager.flush();
     }
-
-//    @Test
-//    @DisplayName("Workbook 저장 - 실패, 이름이 최대 길이를 초과")
-//    void saveWithLongName() {
-//        // given
-//        Workbook workbook = Workbook.builder()
-//                .name("abcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyz")
-//                .isDeleted(false)
-//                .build();
-//
-//        // when, then
-//        assertThatThrownBy(() -> workbookRepository.save(workbook))
-//                .isInstanceOf(DataIntegrityViolationException.class);
-//    }
 
     @Test
     @DisplayName("Workbook id로 조회 - 성공")
